pages/requirements.txt
altair==5.2.0
apturl==0.5.2
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
bcrypt==3.2.0
blinker==1.4
branca==0.7.0
Brlapi==0.8.3
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
comm==0.2.0
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
DateTime==5.4
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==5.0
duplicity==0.8.21
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.0
folium==0.15.1
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
haversine==2.8.0
httplib2==0.20.2
idna==3.3
image==1.5.33
importlib-metadata==4.6.4
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
packaging==23.2
pandas==2.1.4
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pycairo==1.20.1
pycups==2.0.1
pydeck==0.8.1b0
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.2
referencing==0.32.0
reportlab==3.6.8
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.29.0
streamlit-folium==0.17.4
systemd-python==234
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
traitlets==5.14.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
wcwidth==0.2.12
xdg==5
xkit==0.0.0
xyzservices==2023.10.1
zipp==1.0.0
zope.interface==6.1
